Spring Data JPA [#505127]

+++ src/main/resources/db/scripts/007_create_subscriptions_table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE subscriptions (
-    id SERIAL PRIMARY KEY,
-    subscriber_id INT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
-    subscribed_to_id INT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+    id BIGSERIAL PRIMARY KEY,
+    subscriber_id BIGINT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+    subscribed_to_id BIGINT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     UNIQUE (subscriber_id, subscribed_to_id)
 );
